[apiConformance] Rename exec_graph_info and exec_network_base
