+++ modules/arm_plugin/tests/functional/shared_tests_instances/behavior/ov_compiled_model/core_integration.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "behavior/ov_executable_network/get_metric.hpp"
+#include "behavior/ov_compiled_model/get_metric.hpp"
 #include "openvino/runtime/core.hpp"
 
 using namespace ov::test::behavior;
